requirements.txt
pandas
requests
discord.py
python-dotenv
